% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_datasets.R
\name{list_datasets}
\alias{list_datasets}
\title{This function returns a list of all CDRC datasets that can be retrieved with this package. This function takes no arguments.}
\usage{
list_datasets()
}
\value{
A dataframe of the datasets available to request with the API.
}
\description{
This function returns a list of all CDRC datasets that can be retrieved with this package. This function takes no arguments.
}
\examples{
\dontrun{
dataset_list <- list_data()
View(dataset_list)
}
}
