% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listCDRC.R
\name{listCDRC}
\alias{listCDRC}
\title{This function returns a list of all CDRC datasets that can be retrieved with this package. This function takes no arguments.}
\usage{
listCDRC()
}
\value{
A dataframe of the datasets available to request with the API.
}
\description{
This function returns a list of all CDRC datasets that can be retrieved with this package. This function takes no arguments.
}
\examples{
\dontrun{
dataset_list <- listCDRC()
View(dataset_list)
}
}
