% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liverpool}
\alias{liverpool}
\title{Liverpool LSOA boundaries}
\format{
A simple feature object with 297 rows and 7 variables:
\describe{
\item{OBJECTID}{Geographical Feature ID}
\item{LSOA11CD}{LSOA Codes}
\item{LSOA11NM}{LSOA Names}
\item{LSOA11NMW}{LSOA Names in Welsh}
\item{Shape_Are}{LSOA Area}
\item{Shape_Len}{LSOA Length}
\item{geom}{Simple Feature Geometry}
}
}
\source{
\url{https://geoportal.statistics.gov.uk/}
}
\usage{
liverpool
}
\description{
A dataset containing Lower Layer Super Output Areas in Liverpool Local Authority
}
\keyword{datasets}
